pyproject.toml
[tool.poetry]
name = "web3auth-demo"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"
packages = [{include = "web3auth-demo"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "3"
djangorestframework = "^3.15.1"
black = "^24.4.2"
isort = "^5.13.2"
py-ecc = "^7.0.1"
web3 = "6.17.1"
redis = "^5.0.4"
djangorestframework-simplejwt = "5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
